refactor(api): move logging initialization into api crate
Move process-level configuration and logging setup into the top-level
carbide-api crate, leaving api-core responsible for the core service
runtime.

Introduce a dedicated bootstrap interface for the values that cross
the crate boundary, and move logging-only dependencies and tests with
their implementation.
