+++ crates/api/Cargo.toml
@@ -39,12 +39,26 @@ path = "src/main.rs"
 carbide-api-core = { path = "../api-core", default-features = false }
 carbide-api-db = { path = "../api-db", default-features = false }
 carbide-api-web = { path = "../api-web" }
+carbide-instrument = { path = "../instrument" }
+carbide-machine-controller = { path = "../machine-controller" }
 carbide-secrets = { path = "../secrets" }
 carbide-version = { path = "../version" }
+logfmt = { path = "../logfmt" }
+spancounter = { path = "../spancounter" }
+sqlx-query-tracing = { path = "../sqlx-query-tracing" }
 
 # External dependencies. PLEASE KEEP ALPHABETIZED ORDER.
 clap = { workspace = true }
 eyre = { workspace = true }
+ipnetwork = { workspace = true }
+opentelemetry = { workspace = true, features = ["logs"] }
+opentelemetry-otlp = { workspace = true, features = ["grpc-tonic"] }
+opentelemetry_sdk = { workspace = true, features = [
+  "logs",
+  "rt-tokio",
+  "spec_unstable_metrics_views",
+  "testing",
+] }
 sqlx = { workspace = true, features = [
   "runtime-tokio",
   "tls-rustls",
@@ -54,6 +68,11 @@ sqlx = { workspace = true, features = [
 tokio = { workspace = true }
 tokio-util = { workspace = true }
 tracing = { workspace = true }
+tracing-opentelemetry = { workspace = true }
+tracing-subscriber = { workspace = true, features = [
+  "env-filter",
+  "local-time",
+] }
 
 [features]
 default = ["linux-build"]
@@ -61,6 +80,13 @@ linux-build = ["carbide-api-core/linux-build"]
 
 [build-dependencies]
 carbide-version = { path = "../version" }
+
+[dev-dependencies]
+# Internal dependencies. PLEASE KEEP ALPHABETIZED ORDER.
+trace-propagation = { path = "../trace-propagation" }
+
+# External dependencies. PLEASE KEEP ALPHABETIZED ORDER.
+http = { workspace = true }
 
 [lints]
 workspace = true
